added latest stateChangeDateTime to API spec so it is included in events, removed some obsolete calls to eventStore
