+++ .circleci/config.yml
@@ -32,7 +32,7 @@ jobs:
             - ~/.gradle
           key: v1-dependencies-{{ checksum "build.gradle" }}
 
-      - run: ./gradlew test --info
+      - run: ./gradlew test
 
       - run: ./gradlew codeCoverageReport
