+++ impl/src/main/kotlin/nl/nuts/consent/bridge/conversion/CordappToBridgeType.kt
@@ -173,7 +173,8 @@ class CordappToBridgeType {
                 consentRecords = consentRecords,
                 initiatingLegalEntity = source.initiatingLegalEntity,
                 initiatingNode = source.initiatingNode,
-                requestDateTime = source.branchTime
+                requestDateTime = source.branchTime,
+                latestUpdate = source.stateTime
             )
         }
     }
