+++ impl/src/main/kotlin/nl/nuts/consent/bridge/pipelines/ConsentStateChangeToNatsPipeline.kt
@@ -19,7 +19,9 @@
 package nl.nuts.consent.bridge.pipelines
 
 import net.corda.core.contracts.StateAndRef
+import nl.nuts.consent.bridge.EventStoreProperties
 import nl.nuts.consent.bridge.Serialization
+import nl.nuts.consent.bridge.events.apis.EventApi
 import nl.nuts.consent.bridge.events.infrastructure.ClientException
 import nl.nuts.consent.bridge.nats.Event
 import nl.nuts.consent.bridge.nats.EventName
@@ -27,14 +29,26 @@ import nl.nuts.consent.bridge.nats.NATS_CONSENT_REQUEST_SUBJECT
 import nl.nuts.consent.state.ConsentState
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
+import java.util.*
+import javax.annotation.PostConstruct
 
 /**
  * Pipeline to act on produced Nuts ConsentStates
  */
 @Service
 class ConsentStateChangeToNatsPipeline : CordaStateChangeToNatsPipeline<ConsentState>() {
+    @Autowired
+    lateinit var eventstoreProperties: EventStoreProperties
+    lateinit var eventApi: EventApi
+
     private val logger: Logger = LoggerFactory.getLogger(this::class.java)
+
+    @PostConstruct
+    fun initApi() {
+        eventApi = EventApi(eventstoreProperties.url)
+    }
 
     override fun name(): String {
         return stateClass().simpleName
@@ -70,12 +84,12 @@ class ConsentStateChangeToNatsPipeline : CordaStateChangeToNatsPipeline<ConsentS
             // nop
         }
 
-        if (knownEvent == null) {
-            logger.error("Can't find event in event store with UUID: $consentBranchUUID")
-            return
+        if (knownEvent != null) {
+            event.UUID = knownEvent.UUID
+        } else {
+            logger.error("Can't find event in event store with UUID: $consentBranchUUID, generating new UUID")
         }
 
-        event.UUID = knownEvent.UUID
         event.name = EventName.EventConsentDistributed
 
         val jsonBytes = Serialization.objectMapper().writeValueAsBytes(event)
@@ -84,5 +98,22 @@ class ConsentStateChangeToNatsPipeline : CordaStateChangeToNatsPipeline<ConsentS
 
     override fun stateConsumed(stateAndRef: StateAndRef<ConsentState>) {
         logger.debug("Received consentState consumed event from Corda: ${stateAndRef.state.data}")
+    }
+
+    private fun remoteEvent(stateUUID: UUID): Event {
+        return eventToEvent(eventApi.getEvent(stateUUID))
+    }
+
+    private fun eventToEvent(source: nl.nuts.consent.bridge.events.models.Event): Event {
+        return Event(
+            UUID = source.uuid,
+            payload = source.payload,
+            initiatorLegalEntity = source.initiatorLegalEntity,
+            externalId = source.externalId,
+            consentId = source.consentId.toString(),
+            retryCount = source.retryCount,
+            error = source.error,
+            name = EventName.fromString(source.name.value)
+        )
     }
 }
